refactor(eve): keep each session-inbox wire version in its own module
Move the unversioned-era migration and defensive legacy checks to
session-inbox-wire.v0.ts, and the complete v1 schema, inferred type, and
validated v1 encoder to session-inbox-wire.v1.ts. The stable encoder facade
selects the current version while the workflow-safe decoder owns only the
migration chain and normalization, importing v1 as a type.

This makes version history append-only and independently reviewable: v2 is a
new module plus one migration-chain entry rather than an edit to a shared
version implementation.
